backend/requirements.txt
# Core web framework and server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Agent framework
uagents>=0.4.0

# Data models and validation
pydantic>=2.0.0

# HTTP client for API calls
aiohttp>=3.8.0

# PDF processing
PyPDF2>=3.0.0

# Database client
supabase>=2.0.0

# Environment variables
python-dotenv>=1.0.0